task-01-interactive-training/telegram-bot/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
notion-client==2.0.0